app/requirements.txt
feedparser==6.0.11
requests==2.32.3
openai==1.54.0
schedule==1.2.2
python-dotenv==1.0.1
groq==1.0.0
twilio==9.3.5
pytest==8.3.3